+++ src/Symfony/Component/Cache/CacheInterface.php
@@ -1,0 +1,37 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Cache;
+
+use Psr\Cache\CacheItemInterface;
+
+/**
+ * Gets and stores items from a cache.
+ *
+ * On cache misses, a callback is called that should return the missing value.
+ * It is given two arguments:
+ * - the missing cache key
+ * - the corresponding PSR-6 CacheItemInterface object,
+ *   allowing time-based expiration control.
+ *
+ * If you need tag-based invalidation, use TaggableCacheInterface instead.
+ *
+ * @author Nicolas Grekas <p@tchwork.com>
+ */
+interface CacheInterface
+{
+    /**
+     * @param callable(CacheItemInterface):mixed $callback Should return the computed value for the given key/item
+     *
+     * @return mixed The value corresponding to the provided key
+     */
+    public function get(string $key, callable $callback);
+}