+++ src/Symfony/Component/Cache/Tests/Adapter/PhpArrayAdapterTest.php
@@ -21,6 +21,7 @@ use Symfony\Component\Cache\Adapter\PhpArrayAdapter;
 class PhpArrayAdapterTest extends AdapterTestCase
 {
     protected $skippedTests = array(
+        'testGet' => 'PhpArrayAdapter is read-only.',
         'testBasicUsage' => 'PhpArrayAdapter is read-only.',
         'testBasicUsageWithLongKey' => 'PhpArrayAdapter is read-only.',
         'testClear' => 'PhpArrayAdapter is read-only.',